+++ tests/phpunit/unit/Twig/Runtime/ArrayHandlerTest.php
@@ -1,12 +1,12 @@
 <?php
 
-namespace Bolt\Tests\Twig;
+namespace Bolt\Tests\Twig\Runtime;
 
 use Bolt\Tests\BoltUnitTest;
-use Bolt\Twig\Handler\ArrayHandler;
+use Bolt\Twig\Runtime\ArrayRuntime;
 
 /**
- * Class to test Bolt\Twig\Handler\ArrayHandler
+ * Class to test Bolt\Twig\Runtime\ArrayHandler
  *
  * @author Gawain Lynch <gawain.lynch@gmail.com>
  */
@@ -23,7 +23,7 @@ class ArrayHandlerTest extends BoltUnitTest
     public function testOrderEmpty()
     {
         $app = $this->getApp();
-        $handler = new ArrayHandler($app);
+        $handler = new ArrayRuntime($app);
 
         $result = $handler->order([], 'title');
         $this->assertSame([], $result);
@@ -38,7 +38,7 @@ class ArrayHandlerTest extends BoltUnitTest
             ['name' => 'Wayne', 'type' => 'batman'],
         ];
 
-        $handler = new ArrayHandler($app);
+        $handler = new ArrayRuntime($app);
 
         $result = $handler->order($srcArr, 'name', null);
         $this->assertSame('{"1":{"name":"Bruce","type":"clippy"},"0":{"name":"Johno","type":"koala"},"2":{"name":"Wayne","type":"batman"}}', json_encode($result));
@@ -53,7 +53,7 @@ class ArrayHandlerTest extends BoltUnitTest
             ['name' => 'Johno', 'type' => 'koala'],
         ];
 
-        $handler = new ArrayHandler($app);
+        $handler = new ArrayRuntime($app);
 
         $result = $handler->order($srcArr, '-name', null);
         $this->assertSame('{"0":{"name":"Wayne","type":"batman"},"2":{"name":"Johno","type":"koala"},"1":{"name":"Bruce","type":"clippy"}}', json_encode($result));
@@ -68,7 +68,7 @@ class ArrayHandlerTest extends BoltUnitTest
             ['name' => 'Johno', 'type' => 'batman'],
         ];
 
-        $handler = new ArrayHandler($app);
+        $handler = new ArrayRuntime($app);
 
         $result = $handler->order($srcArr, 'name', 'type');
         $this->assertSame('{"1":{"name":"Bruce","type":"clippy"},"2":{"name":"Johno","type":"batman"},"0":{"name":"Johno","type":"koala"}}', json_encode($result));
@@ -83,7 +83,7 @@ class ArrayHandlerTest extends BoltUnitTest
             ['name' => 'Bruce', 'type' => 'clippy'],
         ];
 
-        $handler = new ArrayHandler($app);
+        $handler = new ArrayRuntime($app);
 
         $result = $handler->order($srcArr, 'name', '-type');
         $this->assertSame('{"2":{"name":"Bruce","type":"clippy"},"1":{"name":"Johno","type":"koala"},"0":{"name":"Johno","type":"batman"}}', json_encode($result));
@@ -98,7 +98,7 @@ class ArrayHandlerTest extends BoltUnitTest
             ['name' => 'Bruce', 'type' => 'clippy'],
         ];
 
-        $handler = new ArrayHandler($app);
+        $handler = new ArrayRuntime($app);
 
         $result = $handler->order($srcArr, 'name', '-type');
         $this->assertRegExp('#{"[0-2]":{"name":"Bruce","type":"clippy"},"[0-2]":{"name":"Johno","type":"batman"},"[0-2]":{"name":"Johno","type":"batman"}}#', json_encode($result));
@@ -113,7 +113,7 @@ class ArrayHandlerTest extends BoltUnitTest
             ['name' => 'Johno', 'type' => 'batman'],
         ];
 
-        $handler = new ArrayHandler($app);
+        $handler = new ArrayRuntime($app);
 
         $result = $handler->order($srcArr, 'name');
         $this->assertRegExp('#{"[0-2]":{"name":"Bruce","type":"clippy"},"[0-2]":{"name":"Johno","type":"(batman|koala)"},"[0-2]":{"name":"Johno","type":"(koala|batman)"}}#', json_encode($result));
@@ -122,7 +122,7 @@ class ArrayHandlerTest extends BoltUnitTest
     public function testShuffleString()
     {
         $app = $this->getApp();
-        $handler = new ArrayHandler($app);
+        $handler = new ArrayRuntime($app);
 
         $result = $handler->shuffle('shuffleboard');
         $this->assertSame('shuffleboard', $result);
@@ -140,7 +140,7 @@ class ArrayHandlerTest extends BoltUnitTest
             ->method('shuffle')
         ;
 
-        $handler = new ArrayHandler($app);
+        $handler = new ArrayRuntime($app);
         $handler->shuffle(['shuffle', 'board']);
     }
 }
