+++ tests/phpunit/unit/Twig/Runtime/ImageHandlerTest.php
@@ -1,15 +1,15 @@
 <?php
 
-namespace Bolt\Tests\Twig;
+namespace Bolt\Tests\Twig\Runtime;
 
 use Bolt\Filesystem\Handler\Image;
 use Bolt\Filesystem\Handler\ImageInterface;
 use Bolt\Tests\BoltUnitTest;
-use Bolt\Twig\Handler\ImageHandler;
+use Bolt\Twig\Runtime\ImageRuntime;
 use Symfony\Component\Filesystem\Filesystem;
 
 /**
- * Class to test Bolt\Twig\Handler\ImageHandler
+ * Class to test Bolt\Twig\Runtime\ImageHandler
  *
  * @author Gawain Lynch <gawain.lynch@gmail.com>
  */
@@ -26,7 +26,7 @@ class ImageHandlerTest extends BoltUnitTest
     public function testImageFileNameLegacy()
     {
         $app = $this->getApp();
-        $handler = new ImageHandler($app);
+        $handler = new ImageRuntime($app);
 
         $result = $handler->image('generic-logo.png');
         $this->assertStringStartsWith('/files/generic-logo.png', $result);
@@ -35,7 +35,7 @@ class ImageHandlerTest extends BoltUnitTest
     public function testImageFileNameArray()
     {
         $app = $this->getApp();
-        $handler = new ImageHandler($app);
+        $handler = new ImageRuntime($app);
 
         $result = $handler->image(['filename' => 'generic-logo.png']);
         $this->assertStringStartsWith('/files/generic-logo.png', $result);
@@ -44,7 +44,7 @@ class ImageHandlerTest extends BoltUnitTest
     public function testImageFileNameWidthOnly()
     {
         $app = $this->getApp();
-        $handler = new ImageHandler($app);
+        $handler = new ImageRuntime($app);
 
         $result = $handler->image(['filename' => 'generic-logo.png'], 20);
         $this->assertSame('/thumbs/20x120c/generic-logo.png', $result);
@@ -53,7 +53,7 @@ class ImageHandlerTest extends BoltUnitTest
     public function testImageFileNameHeightOnly()
     {
         $app = $this->getApp();
-        $handler = new ImageHandler($app);
+        $handler = new ImageRuntime($app);
 
         $result = $handler->image(['filename' => 'generic-logo.png'], '', 20);
         $this->assertSame('/thumbs/160x20c/generic-logo.png', $result);
@@ -62,7 +62,7 @@ class ImageHandlerTest extends BoltUnitTest
     public function testImageFileNameWidthHeight()
     {
         $app = $this->getApp();
-        $handler = new ImageHandler($app);
+        $handler = new ImageRuntime($app);
 
         $result = $handler->image(['filename' => 'generic-logo.png'], 20, 20);
         $this->assertSame('/thumbs/20x20c/generic-logo.png', $result);
@@ -71,7 +71,7 @@ class ImageHandlerTest extends BoltUnitTest
     public function testImageFileNameCrop()
     {
         $app = $this->getApp();
-        $handler = new ImageHandler($app);
+        $handler = new ImageRuntime($app);
 
         $result = $handler->image(['filename' => 'generic-logo.png'], 20, 20, 'f');
         $this->assertSame('/thumbs/20x20f/generic-logo.png', $result);
@@ -97,7 +97,7 @@ class ImageHandlerTest extends BoltUnitTest
     public function testImageInfoNotReadable()
     {
         $app = $this->getApp();
-        $handler = new ImageHandler($app);
+        $handler = new ImageRuntime($app);
 
         $image = $handler->imageInfo('koala.jpg', false);
         $this->assertInstanceOf(ImageInterface::class, $image);
@@ -107,7 +107,7 @@ class ImageHandlerTest extends BoltUnitTest
     public function testImageInfo()
     {
         $app = $this->getApp();
-        $handler = new ImageHandler($app);
+        $handler = new ImageRuntime($app);
 
         $image = $handler->imageInfo('generic-logo.png', false);
         $this->assertInstanceOf(ImageInterface::class, $image);
@@ -117,7 +117,7 @@ class ImageHandlerTest extends BoltUnitTest
     public function testPopupEmptyFileName()
     {
         $app = $this->getApp();
-        $handler = new ImageHandler($app);
+        $handler = new ImageRuntime($app);
 
         $result = $handler->popup();
         $this->assertSame('', $result);
@@ -126,7 +126,7 @@ class ImageHandlerTest extends BoltUnitTest
     public function testPopupFileNameOnly()
     {
         $app = $this->getApp();
-        $handler = new ImageHandler($app);
+        $handler = new ImageRuntime($app);
 
         $result = $handler->popup('generic-logo.png');
         $this->assertSame('<a href="/thumbs/1000x750r/generic-logo.png" class="magnific" title="Image: generic-logo.png"><img src="/thumbs/160x120c/generic-logo.png" width="160" height="120" alt="Image: generic-logo.png"></a>', $result);
@@ -135,7 +135,7 @@ class ImageHandlerTest extends BoltUnitTest
     public function testPopupWidth()
     {
         $app = $this->getApp();
-        $handler = new ImageHandler($app);
+        $handler = new ImageRuntime($app);
 
         $result = $handler->popup('generic-logo.png', 50);
         $this->assertSame('<a href="/thumbs/1000x750r/generic-logo.png" class="magnific" title="Image: generic-logo.png"><img src="/thumbs/50x120c/generic-logo.png" width="50" height="120" alt="Image: generic-logo.png"></a>', $result);
@@ -144,7 +144,7 @@ class ImageHandlerTest extends BoltUnitTest
     public function testPopupHeight()
     {
         $app = $this->getApp();
-        $handler = new ImageHandler($app);
+        $handler = new ImageRuntime($app);
 
         $result = $handler->popup('generic-logo.png', null, 50);
         $this->assertSame('<a href="/thumbs/1000x750r/generic-logo.png" class="magnific" title="Image: generic-logo.png"><img src="/thumbs/160x50c/generic-logo.png" width="160" height="50" alt="Image: generic-logo.png"></a>', $result);
@@ -153,7 +153,7 @@ class ImageHandlerTest extends BoltUnitTest
     public function testPopupCrop()
     {
         $app = $this->getApp();
-        $handler = new ImageHandler($app);
+        $handler = new ImageRuntime($app);
 
         $result = $handler->popup('generic-logo.png', null, null, 'f');
         $this->assertSame('<a href="/thumbs/1000x750r/generic-logo.png" class="magnific" title="Image: generic-logo.png"><img src="/thumbs/160x120f/generic-logo.png" width="160" height="120" alt="Image: generic-logo.png"></a>', $result);
@@ -179,7 +179,7 @@ class ImageHandlerTest extends BoltUnitTest
     public function testPopupFileNameArrayWithTitle()
     {
         $app = $this->getApp();
-        $handler = new ImageHandler($app);
+        $handler = new ImageRuntime($app);
 
         $result = $handler->popup(['title' => 'Koala', 'filename' => 'generic-logo.png'], null, null, null, null);
         $this->assertSame('<a href="/thumbs/1000x750r/generic-logo.png" class="magnific" title="Koala"><img src="/thumbs/160x120c/generic-logo.png" width="160" height="120" alt="Koala"></a>', $result);
@@ -188,7 +188,7 @@ class ImageHandlerTest extends BoltUnitTest
     public function testPopupFileNameArrayWithoutTitle()
     {
         $app = $this->getApp();
-        $handler = new ImageHandler($app);
+        $handler = new ImageRuntime($app);
 
         $result = $handler->popup(['filename' => 'generic-logo.png'], null, null, null, null);
         $this->assertSame('<a href="/thumbs/1000x750r/generic-logo.png" class="magnific" title="Image: generic-logo.png"><img src="/thumbs/160x120c/generic-logo.png" width="160" height="120" alt="Image: generic-logo.png"></a>', $result);
@@ -197,7 +197,7 @@ class ImageHandlerTest extends BoltUnitTest
     public function testPopupFileNameArrayWithAlt()
     {
         $app = $this->getApp();
-        $handler = new ImageHandler($app);
+        $handler = new ImageRuntime($app);
 
         $result = $handler->popup(['title' => 'Koala', 'alt' => 'Gum Leaves', 'filename' => 'generic-logo.png'], null, null, null, null);
         $this->assertSame('<a href="/thumbs/1000x750r/generic-logo.png" class="magnific" title="Koala"><img src="/thumbs/160x120c/generic-logo.png" width="160" height="120" alt="Gum Leaves"></a>', $result);
@@ -206,7 +206,7 @@ class ImageHandlerTest extends BoltUnitTest
     public function testShowImageEmptyFileName()
     {
         $app = $this->getApp();
-        $handler = new ImageHandler($app);
+        $handler = new ImageRuntime($app);
 
         $result = $handler->showImage();
         $this->assertSame('', $result);
@@ -215,7 +215,7 @@ class ImageHandlerTest extends BoltUnitTest
     public function testShowImageFileNameOnly()
     {
         $app = $this->getApp();
-        $handler = new ImageHandler($app);
+        $handler = new ImageRuntime($app);
 
         $result = $handler->showImage('generic-logo.png');
         $this->assertSame('<img src="/thumbs/1000x750c/generic-logo.png" width="1000" height="750" alt="">', $result);
@@ -224,7 +224,7 @@ class ImageHandlerTest extends BoltUnitTest
     public function testShowImageWidth()
     {
         $app = $this->getApp();
-        $handler = new ImageHandler($app);
+        $handler = new ImageRuntime($app);
 
         $result = $handler->showImage('generic-logo.png', 50);
         $this->assertSame('<img src="/thumbs/50x28c/generic-logo.png" width="50" height="28" alt="">', $result);
@@ -233,7 +233,7 @@ class ImageHandlerTest extends BoltUnitTest
     public function testShowImageHeight()
     {
         $app = $this->getApp();
-        $handler = new ImageHandler($app);
+        $handler = new ImageRuntime($app);
 
         $result = $handler->showImage('generic-logo.png', null, 50);
         $this->assertSame('<img src="/thumbs/89x50c/generic-logo.png" width="89" height="50" alt="">', $result);
@@ -242,7 +242,7 @@ class ImageHandlerTest extends BoltUnitTest
     public function testShowImageCrop()
     {
         $app = $this->getApp();
-        $handler = new ImageHandler($app);
+        $handler = new ImageRuntime($app);
 
         $result = $handler->showImage('generic-logo.png', null, null, 'f');
         $this->assertSame('<img src="/thumbs/1000x750f/generic-logo.png" width="1000" height="750" alt="">', $result);
@@ -268,7 +268,7 @@ class ImageHandlerTest extends BoltUnitTest
     public function testShowImageFileNameArrayWithTitle()
     {
         $app = $this->getApp();
-        $handler = new ImageHandler($app);
+        $handler = new ImageRuntime($app);
 
         $result = $handler->showImage(['title' => 'Koala', 'filename' => 'generic-logo.png'], null, null, null, null);
         $this->assertSame('<img src="/thumbs/1000x750c/generic-logo.png" width="1000" height="750" alt="Koala">', $result);
@@ -277,7 +277,7 @@ class ImageHandlerTest extends BoltUnitTest
     public function testShowImageFileNameArrayWithoutTitle()
     {
         $app = $this->getApp();
-        $handler = new ImageHandler($app);
+        $handler = new ImageRuntime($app);
 
         $result = $handler->showImage(['filename' => 'generic-logo.png'], null, null, null, null);
         $this->assertSame('<img src="/thumbs/1000x750c/generic-logo.png" width="1000" height="750" alt="">', $result);
@@ -286,7 +286,7 @@ class ImageHandlerTest extends BoltUnitTest
     public function testShowImageFileNameArrayWithAlt()
     {
         $app = $this->getApp();
-        $handler = new ImageHandler($app);
+        $handler = new ImageRuntime($app);
 
         $result = $handler->showImage(['title' => 'Koala', 'alt' => 'Gum Leaves', 'filename' => 'generic-logo.png'], null, null, null, null);
         $this->assertSame('<img src="/thumbs/1000x750c/generic-logo.png" width="1000" height="750" alt="Gum Leaves">', $result);
