+++ tests/phpunit/unit/Twig/Runtime/UtilsHandlerTest.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace Bolt\Tests\Twig;
+namespace Bolt\Tests\Twig\Runtime;
 
 use Bolt\Application;
 use Bolt\Tests\BoltUnitTest;
-use Bolt\Twig\Handler\UtilsHandler;
+use Bolt\Twig\Runtime\UtilsRuntime;
 use Symfony\Component\HttpFoundation\Request;
 
 /**
@@ -31,7 +31,7 @@ class UtilsHandlerTest extends BoltUnitTest
     public function testFileExists()
     {
         $app = $this->getApp();
-        $handler = new UtilsHandler($app);
+        $handler = new UtilsRuntime($app);
 
         $result = $handler->fileExists(__FILE__, false);
         $this->assertTrue($result);
@@ -40,7 +40,7 @@ class UtilsHandlerTest extends BoltUnitTest
     public function testFileExistsSafe()
     {
         $app = $this->getApp();
-        $handler = new UtilsHandler($app);
+        $handler = new UtilsRuntime($app);
 
         $result = $handler->fileExists(__FILE__, true);
         $this->assertFalse($result);
@@ -51,7 +51,7 @@ class UtilsHandlerTest extends BoltUnitTest
         $app = $this->getApp();
         $this->stubVarDumper($app);
         $app['debug'] = true;
-        $handler = new UtilsHandler($app);
+        $handler = new UtilsRuntime($app);
 
         $result = $handler->printBacktrace(5, true);
         $this->assertNull($result);
@@ -62,7 +62,7 @@ class UtilsHandlerTest extends BoltUnitTest
         $app = $this->getApp();
         $this->stubVarDumper($app);
         $app['debug'] = false;
-        $handler = new UtilsHandler($app);
+        $handler = new UtilsRuntime($app);
 
         $result = $handler->printBacktrace(5, false);
         $this->assertNull($result);
@@ -75,7 +75,7 @@ class UtilsHandlerTest extends BoltUnitTest
         $this->stubVarDumper($app);
         $app['debug'] = true;
         $app['config']->set('general/debug_show_loggedoff', false);
-        $handler = new UtilsHandler($app);
+        $handler = new UtilsRuntime($app);
 
         $result = $handler->printBacktrace(5, false);
         $this->assertNull($result);
@@ -89,7 +89,7 @@ class UtilsHandlerTest extends BoltUnitTest
         $app['debug'] = true;
         $app['config']->set('general/debug_show_loggedoff', true);
 
-        $handler = new UtilsHandler($app);
+        $handler = new UtilsRuntime($app);
 
         $result = $handler->printBacktrace(5, false);
         $this->assertCount(5, $result);
@@ -103,7 +103,7 @@ class UtilsHandlerTest extends BoltUnitTest
     {
         $app = $this->getApp();
         $app['debug'] = true;
-        $handler = new UtilsHandler($app);
+        $handler = new UtilsRuntime($app);
 
         $result = $handler->printFirebug(['koala', 'clippy'], 'Danger Detected!', true);
         $this->assertNull($result);
@@ -113,7 +113,7 @@ class UtilsHandlerTest extends BoltUnitTest
     {
         $app = $this->getApp();
         $app['debug'] = false;
-        $handler = new UtilsHandler($app);
+        $handler = new UtilsRuntime($app);
 
         $result = $handler->printFirebug(['koala', 'clippy'], 'Danger Detected!', true);
         $this->assertNull($result);
@@ -123,7 +123,7 @@ class UtilsHandlerTest extends BoltUnitTest
     {
         $app = $this->getApp();
         $app['debug'] = true;
-        $handler = new UtilsHandler($app);
+        $handler = new UtilsRuntime($app);
 
         $result = $handler->printFirebug(['koala', 'clippy'], 'Danger Detected!', true);
         $this->assertNull($result);
@@ -142,7 +142,7 @@ class UtilsHandlerTest extends BoltUnitTest
         ->method('info');
         $app['logger.firebug'] = $logger;
 
-        $handler = new UtilsHandler($app);
+        $handler = new UtilsRuntime($app);
 
         $handler->printFirebug(['koala', 'clippy'], 'Danger Detected!', false);
     }
@@ -159,7 +159,7 @@ class UtilsHandlerTest extends BoltUnitTest
             ->method('info');
         $app['logger.firebug'] = $logger;
 
-        $handler = new UtilsHandler($app);
+        $handler = new UtilsRuntime($app);
 
         $handler->printFirebug('Danger Detected!', ['koala', 'clippy'], false);
     }
@@ -175,7 +175,7 @@ class UtilsHandlerTest extends BoltUnitTest
             ->method('info');
         $app['logger.firebug'] = $logger;
 
-        $handler = new UtilsHandler($app);
+        $handler = new UtilsRuntime($app);
 
         $handler->printFirebug(['koala and clippy'], ['Danger Detected!'], false);
     }
@@ -183,7 +183,7 @@ class UtilsHandlerTest extends BoltUnitTest
     public function testRedirectSafe()
     {
         $app = $this->getApp();
-        $handler = new UtilsHandler($app);
+        $handler = new UtilsRuntime($app);
 
         $result = $handler->redirect('/clippy/koala', true);
         $this->assertNull($result);
@@ -200,7 +200,7 @@ class UtilsHandlerTest extends BoltUnitTest
         }
 
         $app = $this->getApp();
-        $handler = new UtilsHandler($app);
+        $handler = new UtilsRuntime($app);
 
         $this->expectOutputRegex('/Redirecting to/i');
         $handler->redirect('/clippy/koala', false);
@@ -212,7 +212,7 @@ class UtilsHandlerTest extends BoltUnitTest
         $app = $this->getApp();
         $request = Request::createFromGlobals();
         $app['request'] = $request;
-        $handler = new UtilsHandler($app);
+        $handler = new UtilsRuntime($app);
 
         $result = $handler->request('route', 'GET', true, true);
         $this->assertNull($result);
@@ -224,7 +224,7 @@ class UtilsHandlerTest extends BoltUnitTest
         $request = Request::createFromGlobals();
         $request->query->set('koala', 'gum leaves');
         $app['request'] = $request;
-        $handler = new UtilsHandler($app);
+        $handler = new UtilsRuntime($app);
 
         $result = $handler->request('koala', 'GET', true, false);
         $this->assertSame('gum leaves', $result);
@@ -236,7 +236,7 @@ class UtilsHandlerTest extends BoltUnitTest
         $request = Request::createFromGlobals();
         $request->request->set('koala', 'gum leaves');
         $app['request'] = $request;
-        $handler = new UtilsHandler($app);
+        $handler = new UtilsRuntime($app);
 
         $result = $handler->request('koala', 'POST', true, false);
         $this->assertSame('gum leaves', $result);
@@ -248,7 +248,7 @@ class UtilsHandlerTest extends BoltUnitTest
         $request = Request::createFromGlobals();
         $request->attributes->set('koala', 'gum leaves');
         $app['request'] = $request;
-        $handler = new UtilsHandler($app);
+        $handler = new UtilsRuntime($app);
 
         $result = $handler->request('koala', 'PATCH', true, false);
         $this->assertSame('gum leaves', $result);
